requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
debugpy==1.8.9
Django==5.0
django-allauth==65.3.1
django-browser-reload==1.17.0
django-filter==24.3
djangorestframework==3.15.2
Faker==35.2.0
fontawesomefree==6.6.0
idna==3.10
Markdown==3.7
mysql-connector-python==8.3.0
oauthlib==3.2.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-environ==0.4.54
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.3.0
gunicorn
